Add note about WPF GUI
