+++ README.md
@@ -22,3 +22,7 @@ Bei CSV wird `output.csv` erzeugt, bei XLSX `output.xlsx`. Alle Zeilenumbrüche 
 2. Mehrfache Leerzeichen auf ein Leerzeichen reduzieren.
 3. Vor den Begriffen "Anschlüsse:", "Technische Daten:", "Daten:" und "Betriebsspannung:" einen Zeilenumbruch einfügen. Steht "Daten:" alleine, auch danach einen Zeilenumbruch. "Technische Daten" darf nicht getrennt werden. Nach jedem Doppelpunkt einen Zeilenumbruch einfügen. Beginnt danach eine Aufzählung, jede Zeile mit `-` beginnen und zuvor einen Zeilenumbruch einfügen.
 4. Technische Daten logisch und übersichtlich strukturieren. Keine Formatierungen einsetzen, leere Felder bleiben leer und vorhandene Schreibweise wird nicht korrigiert.
+
+## WPF-Anwendung
+
+Neben dem Konsolenprogramm gibt es ein WPF-Projekt. Darüber kann eine CSV- oder XLSX-Datei ausgewählt und per Button die Umwandlung gestartet werden. Die Ausgabedatei wird im selben Ordner gespeichert.
