+++ src/Honeywell.ArticleTransformer/Honeywell.ArticleTransformer.Gui/App.xaml
@@ -1,0 +1,7 @@
+<Application x:Class="Honeywell.ArticleTransformer.Gui.App"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             StartupUri="MainWindow.xaml">
+    <Application.Resources>
+    </Application.Resources>
+</Application>
