+++ src/Honeywell.ArticleTransformer/Honeywell.ArticleTransformer.Gui/App.xaml.cs
@@ -1,0 +1,8 @@
+using System.Windows;
+
+namespace Honeywell.ArticleTransformer.Gui
+{
+    public partial class App : Application
+    {
+    }
+}
