+++ src/Honeywell.ArticleTransformer/Honeywell.ArticleTransformer.Gui/Honeywell.ArticleTransformer.Gui.csproj
@@ -1,0 +1,12 @@
+<Project Sdk="Microsoft.NET.Sdk.WindowsDesktop">
+  <PropertyGroup>
+    <OutputType>WinExe</OutputType>
+    <TargetFramework>net9.0-windows</TargetFramework>
+    <Nullable>enable</Nullable>
+    <UseWPF>true</UseWPF>
+    <EnableWindowsTargeting>true</EnableWindowsTargeting>
+  </PropertyGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Honeywell.ArticleTransformer.Core\Honeywell.ArticleTransformer.Core.csproj" />
+  </ItemGroup>
+</Project>
