+++ src/Honeywell.ArticleTransformer/Honeywell.ArticleTransformer/Honeywell.ArticleTransformer.csproj
@@ -6,4 +6,7 @@
     <Nullable>enable</Nullable>
     <Version>1.0.0</Version>
   </PropertyGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Honeywell.ArticleTransformer.Core\Honeywell.ArticleTransformer.Core.csproj" />
+  </ItemGroup>
 </Project>
